fix: Truncate query in exceptions and add test coverage for protocol limits (#15)
With this change we are checking limits around:
- headers
- request size
- parameter sizes

The driver already passed all limits but the tests uncoverd a
significant inefficiency around exception creation. With multi megabyte
queries this was causing multi second hangs. Thus we now truncate the
query in the exception message.
If the query is truncated it'll have a `<truncated>` suffix