fix:  prevent subsequent getQueryInfo on small query (#41)
Enhancements to query result handling:

*
[`jdbc-core/src/main/java/com/salesforce/datacloud/jdbc/core/listener/AdaptiveQueryStatusListener.java`](diffhunk://#diff-c22f047bc3e8cd467e08d7343f2ca7025eab4aa351895dae3fd4d456615b0620R116-R127):
Added the `allResultsInHead` method to check if all results are produced
and the chunk count is less than 2, and used this method in the `tail`
method to return an empty stream if all results are in the head.

New test case:

*
[`jdbc-core/src/test/java/com/salesforce/datacloud/jdbc/core/partial/DataCloudQueryPollingMockTests.java`](diffhunk://#diff-f6307e3bd1b0c180e5088a11d03bafc751addf629f6ad8059951763bd7a63348R92-R106):
Added the `noFinalGetQueryInfoOnSmallData` test to verify that
`getQueryInfo` is not called when dealing with small data sets.