fix: add missing protobuf-java library and add flattenJars task (#50)
This pull request includes changes to update dependencies, improve
logging, and reorganize the tracing utilities.

* Added `flattenJars` task to easily test with dbeaver without needing
to wait for publication to maven central
* Updated the `revision` property in `gradle.properties` from `0.26.1`
to `0.26.2`.
* Added a new dependency for `protobuf-java` in
`gradle/libs.versions.toml` and included it in the `grpc-impl` bundle.
* Added logging for the `lakehouse` response in the `getLakehouse`
method of `DataCloudTokenProcessor`.
* Introduced `logTimedValue` for measuring execution time in the
`DataspaceClient` class and added trace and span IDs to HTTP headers for
better tracing.
* Moved tracing-related classes (`EncodingUtils`, `TemporaryBuffers`,
`Tracer`, and their tests) from the `jdbc-core` module to a new
`jdbc-util` module under the `com.salesforce.datacloud.jdbc.tracing`
package.