fix: using channel builder in DataCloudConnection should always cleanup (#53)
We realized that allowing someone to opt-out of
`closeChannelWithConnection` when constructing a `DataCloudConnection`
with a `ManagedChannelBuilder` would cause the driver to leak that
connection.

This change removes those overloads and just passes `true` to the one
build method that accepts a `DataCloudJdbcManagedChannel` and
`closeChannelWithConnection`, therefore someone who has a handle to the
channel and knows they opted out of automatic cleanup can clean up their
channel later.