build: add windows to build matrix and caching to gradle hyper tasks (#65)
Since I don't have a windows machine to run this on to fix #56 I figured
it might be a good idea to add a windows target to our build matrix. I
also updated the hyper gradle tasks with inputs/outputs to help gradle
decide how to cache.

### .github/workflows/reusable-build-publish.yml:
This mostly adds a Windows build step, so build and test now run in
parallel on both Linux and Windows. Because of differences in shells
between Linux and Windows, I decided to use a GitHub Action to display
the file tree on failure, rather than writing logic to run ls correctly
on each platform. Otherwise, this file is essentially unchanged.

### buildSrc/src/main/kotlin/hyper-conventions.gradle.kts:
This makes some minor changes to include support for x86_64 macOS, and
properly wires through the subtle differences in the Windows zip
compared to the Linux zip. For example, hyperd.exe is in bin instead of
lib like hyperd is. I also specified which files are inputs and outputs
for these hyper Gradle tasks, so they don't run unnecessarily.

###
jdbc-core/src/test/java/com/salesforce/datacloud/jdbc/core/partial/DataCloudQueryPollingMockTests.java:
This test has been getting flakier. I’m not sure if it’s due to the
recent update to hyper, or if it’s a difference in Windows. In either
case, we should stop relying on hyper to sleep, and instead do the
interface extraction refactor so we can mock an executor and properly
check that retries happen that way.

###
jdbc-core/src/test/java/com/salesforce/datacloud/jdbc/exception/QueryExceptionHandlerTest.java:
This was just broken on Windows because of `\r\n` instead of `\n`.