bug: Fix ResultMetaData for various types (#70)
This commit ensures that the type metadata of the query resultsets is
properly read and exposed through JDBC.
Noteable changes include:

The char type is now represented as such instead of Varchar
decimal precision and scale are properly read
array element types are properly set
timestamp with timezone now have the correct type name
In addition to that the information in various other minor metadata
functions like isSigned, isSearchable, ... is also now aligned with the
standard and cross verified with the Postgres JDBC driver as reference.

The changes are structured the following:

The jdbc-reference module includes a reference value generator. It uses
the PostgreSQL JDBC driver connecting against a true Postgres server to
get reference information for result metadata & data. This information
is then persisted to a file so that it can be used in unit tests
The JDBCTypesTest unit test compares the behavior of the Data Cloud JDBC
driver against the values in the reference
The other changes in jdbc-core are to implement the correct type
behavior.

(Proper discussion happened in PR#69(