fix: apply refactor > delombok to remove @UtilityClass (#73)
I suspect that there's an issue with the way `@UtilityClass` works and
can cause problems with Spark at the very least. It took about 60
seconds to use https://plugins.jetbrains.com/plugin/6317-lombok to
remove all instances of it.

The error we see when using the driver in Spark is
"java.lang.NoClassDefFoundError: Could not initialize class
com.salesforce.datacloud.jdbc.core.QueryMetadataUtil" - I saw
`QueryMetadataUtil` is marked with `@UtilityClass` and recalled there
are some drawbacks to its usage, that combined with it being marked as
"negative" in their
[docs](https://projectlombok.org/features/experimental/UtilityClass)
leads me to think we should remove it in either case, since repro is
relatively simple once inside Spark we can see whether or not it will
fix.